refactor: modularize HomeView into separate component files
- Extract HomeHeaderView into separate file
- Extract CollectionStatusCard into separate file
- Extract CategoryGridView and CategoryCard into separate file
- Extract SeasonalHighlightsCard with related models into separate file
- Extract RecentDonationsCard with DonationItem into separate file
- Extract Color hex extension into ColorExtensions file
- Reduce HomeView.swift from 779 lines to ~80 lines
- Improve code organization and maintainability
